+++ pure_extensions/lib/pure_extensions.dart
@@ -1,13 +1,12 @@
 library pure_extensions;
 
+export 'package:rational/rational.dart';
 export 'package:time/time.dart';
 
 export 'src/dart/collections/iterable.dart';
 export 'src/dart/collections/list.dart';
 export 'src/dart/collections/map.dart';
 export 'src/dart/collections/set.dart';
-export 'src/dart/color/color_extension.dart';
-export 'src/dart/color/hsl_color.dart';
 export 'src/dart/primitives/date_time.dart';
 export 'src/dart/primitives/duration.dart';
 export 'src/dart/primitives/geo_point.dart';